# Add symbolic performance repair