This PR adds a performance repair discovered by our tool for `deepcoderdsl`. It reduces the running time of:

`program1.rkt` from 5s to 4s
`program0.rkt` from 4s to 2s 

It did not negatively impact any other files.

This should not affect the correctness of the program. From a quick glance, the output after applying the repair looks the same.